requirements.txt
fastapi[standard]
passlib[bcrypt]
types-passlib
pyjwt